+++ src/components/data-management/SamplesTable.jsx
@@ -1,10 +1,10 @@
 /* eslint-disable import/no-unresolved */
 /* eslint-disable react/jsx-props-no-spreading */
-import React, { useEffect, useState } from 'react';
+import React, { useEffect, useState, useRef } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import PropTypes from 'prop-types';
 import {
-  Table, Row, Col, Typography,
+  Table, Row, Col, Typography, Space, Button,
 } from 'antd';
 import {
   MenuOutlined,
@@ -14,24 +14,209 @@ import { sortableHandle, sortableContainer, sortableElement } from 'react-sortab
 import { DEFAULT_NA } from 'redux/reducers/projects/initialState';
 import { updateExperiment } from 'redux/actions/experiments';
 import { updateProject } from 'redux/actions/projects';
-
 import { Storage } from 'aws-amplify';
 import UploadStatus from 'utils/upload/UploadStatus';
 import { arrayMoveImmutable } from 'utils/array-move';
 import downloadFromUrl from 'utils/data-management/downloadFromUrl';
-import { UploadCell, SampleNameCell } from './SamplesTableCells';
+import {
+  updateSample,
+} from '../../redux/actions/samples';
+import {
+  deleteMetadataTrack,
+  createMetadataTrack,
+} from '../../redux/actions/projects';
+import { UploadCell, SampleNameCell, EditableFieldCell } from './SamplesTableCells';
+import MetadataEditor from './MetadataEditor';
+import SpeciesSelector from './SpeciesSelector';
+import UploadDetailsModal from './UploadDetailsModal';
+import MetadataPopover from './MetadataPopover';
 
-const { Paragraph } = Typography;
+import MetadataColumn from './MetadataColumn';
+import { metadataNameToKey, metadataKeyToName, temporaryMetadataKey } from '../../utils/data-management/metadataUtils';
+import validateInputs from '../../utils/validateInputs';
+
+const { Paragraph, Text } = Typography;
 
 const SamplesTable = (props) => {
-  const { height, tableColumns } = props;
+  const { height } = props;
   const dispatch = useDispatch();
   const [tableData, setTableData] = useState([]);
+  const [uploadDetailsModalVisible, setUploadDetailsModalVisible] = useState(false);
+
   const projects = useSelector((state) => state.projects);
   const samples = useSelector((state) => state.samples);
   const { activeProjectUuid } = useSelector((state) => state.projects.meta) || false;
   const activeProject = useSelector((state) => state.projects[activeProjectUuid]) || false;
   const environment = useSelector((state) => state?.networkResources?.environment);
+  const uploadDetailsModalDataRef = useRef(null);
+
+  const [sampleNames, setSampleNames] = useState(new Set());
+  const [isAddingMetadata, setIsAddingMetadata] = useState(false);
+
+  const anyProjectsAvailable = projects?.ids?.length;
+
+  const initialTableColumns = [
+    {
+      index: 0,
+      key: 'sort',
+      dataIndex: 'sort',
+      width: 30,
+      render: () => <DragHandle />,
+    },
+    {
+      className: 'data-test-class-sample-cell',
+      index: 1,
+      key: 'sample',
+      title: 'Sample',
+      dataIndex: 'name',
+      fixed: true,
+      render: (text, record, indx) => <SampleNameCell cellInfo={{ text, record, indx }} />,
+    },
+    {
+      index: 2,
+      key: 'barcodes',
+      title: 'barcodes.tsv',
+      dataIndex: 'barcodes',
+      render: (tableCellData) => renderUploadCell('barcodes', tableCellData),
+    },
+    {
+      index: 3,
+      key: 'genes',
+      title: 'genes.tsv',
+      dataIndex: 'genes',
+      render: (tableCellData) => renderUploadCell('genes', tableCellData),
+    },
+    {
+      index: 4,
+      key: 'matrix',
+      title: 'matrix.mtx',
+      dataIndex: 'matrix',
+      render: (tableCellData) => renderUploadCell('matrix', tableCellData),
+    },
+    {
+      index: 5,
+      key: 'species',
+      title: () => (
+        <Space>
+          <Text>Species</Text>
+          <MetadataEditor
+            onReplaceEmpty={(value) => setCells(value, 'species', 'REPLACE_EMPTY')}
+            onReplaceAll={(value) => setCells(value, 'species', 'REPLACE_ALL')}
+            onClearAll={() => setCells(null, 'species', 'CLEAR_ALL')}
+            massEdit
+          >
+            <SpeciesSelector />
+          </MetadataEditor>
+        </Space>
+      ),
+      dataIndex: 'species',
+      render: (organismId, record) => (
+        <SpeciesSelector
+          value={organismId}
+          onChange={(value) => {
+            dispatch(updateSample(record.uuid, { species: value }));
+          }}
+        />
+      ),
+      width: 200,
+    },
+  ];
+
+  const [tableColumns, setTableColumns] = useState(initialTableColumns);
+
+  const validationParams = {
+    existingNames: sampleNames,
+  };
+
+  useEffect(() => {
+    if (activeProject && activeProject.samples.length > 0) {
+      // if there are samples - build the table columns
+      setSampleNames(new Set(activeProject.samples.map((id) => samples[id]?.name.trim())));
+      const metadataColumns = activeProject?.metadataKeys.map(
+        (metadataKey) => createInitializedMetadataColumn(metadataKeyToName(metadataKey)),
+      ) || [];
+      setTableColumns([...initialTableColumns, ...metadataColumns]);
+    } else {
+      setTableColumns([]);
+      setSampleNames(new Set());
+    }
+  }, [samples, activeProject]);
+
+  const deleteMetadataColumn = (name) => {
+    setTableColumns([...tableColumns.filter((entryName) => entryName !== name)]);
+    dispatch(deleteMetadataTrack(name, activeProjectUuid));
+  };
+
+  const createMetadataColumn = () => {
+    setIsAddingMetadata(true);
+
+    const key = temporaryMetadataKey(tableColumns);
+    const metadataColumn = {
+      key,
+      fixed: 'right',
+      title: () => (
+        <MetadataPopover
+          existingMetadata={activeProject.metadataKeys}
+          onCreate={(name) => {
+            const newMetadataColumn = createInitializedMetadataColumn(name);
+
+            setTableColumns([...tableColumns, newMetadataColumn]);
+            dispatch(createMetadataTrack(name, activeProjectUuid));
+
+            setIsAddingMetadata(false);
+          }}
+          onCancel={() => {
+            deleteMetadataColumn(key);
+            setIsAddingMetadata(false);
+          }}
+          message='Provide new metadata track name'
+          visible
+        >
+          <Space>
+            New Metadata Track
+          </Space>
+        </MetadataPopover>
+      ),
+      width: 200,
+    };
+    setTableColumns([...tableColumns, metadataColumn]);
+  };
+
+  const createInitializedMetadataColumn = (name) => {
+    const key = metadataNameToKey(name);
+
+    const newMetadataColumn = {
+      key,
+      title: () => (
+        <MetadataColumn
+          name={name}
+          validateInput={
+            (newName, metadataNameValidation) => validateInputs(
+              newName, metadataNameValidation, validationParams,
+            ).isValid
+          }
+          setCells={setCells}
+          deleteMetadataColumn={deleteMetadataColumn}
+          key={key}
+          activeProjectUuid={activeProjectUuid}
+        />
+      ),
+      width: 200,
+      dataIndex: key,
+      render: (cellValue, record, rowIdx) => (
+        <EditableFieldCell
+          initialText={DEFAULT_NA}
+          cellText={cellValue}
+          dataIndex={key}
+          rowIdx={rowIdx}
+          onAfterSubmit={(newValue) => {
+            dispatch(updateSample(record.uuid, { metadata: { [key]: newValue } }));
+          }}
+        />
+      ),
+    };
+    return newMetadataColumn;
+  };
 
   const DragHandle = sortableHandle(() => <MenuOutlined style={{ cursor: 'grab', color: '#999' }} />);
 
@@ -41,58 +226,52 @@ const SamplesTable = (props) => {
       file,
     } = tableCellData;
     const showDetails = () => {
-      console.warn('implement showing file details');
-      // uploadDetailsModalDataRef.current = {
-      //   sampleUuid,
-      //   fileCategory: columnId,
-      //   file,
-      // };
-      // setUploadDetailsModalVisible(true);
+      uploadDetailsModalDataRef.current = {
+        sampleUuid,
+        fileCategory: columnId,
+        file,
+      };
+      setUploadDetailsModalVisible(true);
     };
     return (
       <UploadCell file={file} showDetails={() => showDetails('barcodes', tableCellData)} />
     );
   };
 
-  // const tableColumns = [
-  //   {
-  //     index: 0,
-  //     key: 'sort',
-  //     dataIndex: 'sort',
-  //     width: 30,
-  //     render: () => <DragHandle />,
-  //   },
-  //   {
-  //     className: 'data-test-class-sample-cell',
-  //     index: 1,
-  //     key: 'sample',
-  //     title: 'Sample',
-  //     dataIndex: 'name',
-  //     fixed: true,
-  //     render: (text, record, indx) => <SampleNameCell cellInfo={{ text, record, indx }} />,
-  //   },
-  //   {
-  //     index: 2,
-  //     key: 'barcodes',
-  //     title: 'barcodes.tsv',
-  //     dataIndex: 'barcodes',
-  //     render: (tableCellData) => renderUploadCell('barcodes', tableCellData),
-  //   },
-  //   {
-  //     index: 3,
-  //     key: 'genes',
-  //     title: 'genes.tsv',
-  //     dataIndex: 'genes',
-  //     render: (tableCellData) => renderUploadCell('genes', tableCellData),
-  //   },
-  //   {
-  //     index: 4,
-  //     key: 'matrix',
-  //     title: 'matrix.mtx',
-  //     dataIndex: 'matrix',
-  //     render: (tableCellData) => renderUploadCell('matrix', tableCellData),
-  //   },
-  // ];
+  const MASS_EDIT_ACTIONS = [
+    'REPLACE_EMPTY',
+    'REPLACE_ALL',
+    'CLEAR_ALL',
+  ];
+
+  const createUpdateObject = (value, metadataKey) => {
+    const updateObject = metadataKey === 'species' ? { species: value } : { metadata: { [metadataKey]: value } };
+    return updateObject;
+  };
+
+  const setCells = (value, metadataKey, actionType) => {
+    if (!MASS_EDIT_ACTIONS.includes(actionType)) return;
+    const updateObject = createUpdateObject(value, metadataKey);
+
+    const canUpdateCell = (sampleUuid, action) => {
+      if (action !== 'REPLACE_EMPTY') return true;
+
+      const isSpeciesEmpty = (uuid) => metadataKey === 'species' && !samples[uuid].species;
+      const isMetadataEmpty = (uuid) => metadataKey !== 'species'
+        && (!samples[uuid].metadata[metadataKey]
+          || samples[uuid].metadata[metadataKey] === DEFAULT_NA);
+
+      return isMetadataEmpty(sampleUuid) || isSpeciesEmpty(sampleUuid);
+    };
+
+    activeProject.samples.forEach(
+      (sampleUuid) => {
+        if (canUpdateCell(sampleUuid, actionType)) {
+          dispatch(updateSample(sampleUuid, updateObject));
+        }
+      },
+    );
+  };
 
   useEffect(() => {
     if (!activeProject || !samples[activeProject.samples[0]]) {
@@ -179,35 +358,55 @@ const SamplesTable = (props) => {
     </Paragraph>
   );
   return (
-    <Row>
-      <Col>
-        <Table
-          size='small'
-          scroll={{
-            x: 'max-content',
-            y: height - 250,
-          }}
-          bordered
-          columns={tableColumns}
-          dataSource={tableData}
-          sticky
-          pagination={false}
-          locale={{ emptyText: noDataText }}
-          components={{
-            body: {
-              wrapper: DragContainer,
-              row: DraggableRow,
-            },
-          }}
-        />
-      </Col>
-    </Row>
+    <>
+      <UploadDetailsModal
+        sampleName={samples[uploadDetailsModalDataRef.current?.sampleUuid]?.name}
+        uploadDetailsModalDataRef={uploadDetailsModalDataRef}
+        visible={uploadDetailsModalVisible}
+        onCancel={() => setUploadDetailsModalVisible(false)}
+      />
+      <Button
+        disabled={
+          !anyProjectsAvailable
+          || activeProject?.samples?.length === 0
+          || isAddingMetadata
+        }
+        onClick={() => {
+          createMetadataColumn();
+        }}
+      >
+        Add metadata
+      </Button>
+      <Row>
+        <Col>
+          <Table
+            size='small'
+            scroll={{
+              x: 'max-content',
+              y: height - 250,
+            }}
+            bordered
+            columns={tableColumns}
+            dataSource={tableData}
+            sticky
+            pagination={false}
+            locale={{ emptyText: noDataText }}
+            components={{
+              body: {
+                wrapper: DragContainer,
+                row: DraggableRow,
+              },
+            }}
+          />
+        </Col>
+      </Row>
+    </>
   );
 };
 
 SamplesTable.propTypes = {
   height: PropTypes.number.isRequired,
-  tableColumns: PropTypes.array.isRequired,
+  // tableColumns: PropTypes.array.isRequired,
 };
 
 export default React.memo(SamplesTable);
